refactor: move sendgrid client into the utils lib
