+++ apps/server/api/src/server/server.ts
@@ -5,9 +5,9 @@ import { setupGlobalMiddleware } from '@ztp/server/middleware';
 import { apolloServer, applyGraphQLEndpoint } from './graphql';
 
 // ZTP_AFTER_CLONE -> delete the below import
-import { applyApiAuthRoutes } from './auth/demo.auth';
+// import { applyApiAuthRoutes } from './auth/demo.auth';
 // ZTP_AFTER_CLONE -> uncomment the below import
-// import { applyApiAuthRoutes } from './auth/auth';
+import { applyApiAuthRoutes } from './auth/auth';
 
 import { config } from '../environments';
 import { connect } from 'mongoose';
