+++ libs/server/auth/src/index.ts
@@ -1,9 +1,11 @@
 export * from './lib/models';
+export * from './lib/graphql';
 export {
   isPasswordAllowed,
   generateAuthGuardConfig,
   generateAuthModuleConfig,
   createPublicPemFromPrivate,
+  createEmailMessage,
 } from './lib/auth-utils';
 export {
   AuthModuleConfig,
@@ -19,9 +21,6 @@ export {
 } from './lib/auth.interface';
 export { getRestGuards } from './lib/routes/router-guards';
 export { applyAuthRoutes } from './lib/routes/auth-routes';
-export { configureSendGridEmail } from './lib/send-email';
-
-export * from './lib/graphql';
 
 /**
  * ZTP_AFTER_CLONE -> Delete from this line onwards after cloning the repo
@@ -29,4 +28,4 @@ export * from './lib/graphql';
  * For further details, see
  * https://github.com/jonathonadams/zero-to-production/tree/master/libs/server/auth/README.md
  */
-export * from './lib/demo/index';
+export * from './lib/demo';
