+++ libs/server/utils/src/lib/email/sendgrid.ts
@@ -1,0 +1,8 @@
+import { setApiKey, send, MailDataRequired } from '@sendgrid/mail';
+
+export function configureSendgrid(apiKey: string) {
+  // Set the api key
+  setApiKey(apiKey);
+
+  return async (msg: MailDataRequired) => send(msg);
+}
